% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrDiagram.R
\name{mrDiagram}
\alias{mrDiagram}
\title{Mediator Plot}
\usage{
mrDiagram(
  fit = NULL,
  method = c("diff", "prod"),
  radx = 0.1,
  rady = 0.04,
  xmargin = 0.01,
  box.col = "white",
  xlim = c(0, 1),
  ylim = NULL
)
}
\arguments{
\item{fit}{fit table}

\item{method}{mediator method either difference or product}
}
\description{
Mediator Plot
}
\examples{
\dontrun{
fit = mrMediator(data=DTC,
exposure = list(name='TSH', type = 'continuous'),
mediator = list(name ='SHBG', type = 'continuous'),
outcome = list(name='DTC', type ='binary'))
mrDiagram(fit =fit, method = 'prod')
}
}
