% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrMediator.R
\name{mrMediator}
\alias{mrMediator}
\title{Perform mediator MR}
\usage{
mrMediator(data, exposure, mediator, outcome)
}
\arguments{
\item{data}{input data}

\item{exposure}{list of name of exposure and type}

\item{mediator}{list of name of mediator and type}

\item{outcome}{list of name of outcome and type}
}
\description{
Perform mediator MR
}
\examples{
fit = mrMediator(data=DTC,exposure = list(name='TSH', type = 'continuous'),mediator = list(name ='SHBG', type = 'continuous'),outcome = list(name='DTC', type ='binary'))
}
