% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrMediator.R
\name{readMRData}
\alias{readMRData}
\title{Read data in MR format}
\usage{
readMRData(data, type, clump = FALSE)
}
\arguments{
\item{data}{input dataset}

\item{type}{whether exposure or outcome}
}
\description{
Read data in MR format
}
